haskell-org-committee: merge in content from the wiki
As outlined in #59, there are two pages which describe the Haskell.org committee and it's roles, responsibilities, process, history, etc. One is here on haskell.org, the other is on the wiki at https://wiki.haskell.org/Haskell.org_committee. Both pages appear authoritative, but differ in their content. At the same time, the general consensus is to move content off the wiki (particularly the more "official" content such as this). We would then replace that wiki article with a link to the this page. This is a first attempt which is more of a "bulk import" (or "merge") of the existing content. Note that there are a few places where the differences in content on the haskell.org page and wiki article may seem to be in conflict (or where one does not go as far as the other). These discrepancies and conflicts should be reviewed and updated. Some of the wiki content is more verbose than that on this page, and may still deserve some word-smithing.

A rough list of changes:
* collected the full list of links at the bottom of the page
* move the list of "committee members" a little lower in the page, bumping contact info and responsibilities above
* Included (merged) process, history and responsibilities content (from the wiki) that was missing here on this page