Simplify Language Report section of Documentation
The source repo is not particularly useful.  Let's not remove it for
now, but giving the `git clone` command seems overly specific.  People
that need to clone the repo typically already know how to do it.

Also merge the paragraphs to make this one short paragraph, otherwise
we're making this marginal interest look more useful than it is.