Delete Alex and happy websites, they should redirect to docs (#228)
(Plan from https://github.com/haskell/alex/issues/215, for example.)

The old websites in addition to being in this repo's history are in
`old-www` branches in both projects' repos. That should be enough
redundancy.

That actual adding of the redirects is done out of band, so this does
the other steps.