Update LYAH link to a working one
As of recent, https://learnyouahaskell.com/chapters lost any CSS / JS because of hard-coded `<base href="http://learnyouahaskell.com/">` (modern browsers are not keen to leave safety of HTTPS). Let's point to https://learnyouahaskell.github.io, which is a community-maintained fork. Besides working CSS, it is more up to date with modern ecosystem.