Updated info about Haskell.org, Inc.
Provides link to current bylaws of the corporation.

Follows more closely the purpose set out in the current bylaws.

Distinguishes the directors of the corporation from the Committee.

Corrects the date of incorporation, based on the information available at https://apps.dos.ny.gov/publicInquiry/.

Adds the DOS ID of the corporation.

Also reformats to avoid long lines.